fix issue in two and three

+++ GO_WS/围棋应用程序/src/eddie/wu/domain/StepMemo.java
@@ -96,7 +96,7 @@ public class StepMemo implements Serializable {
 	/**
 	 * @return Returns the color.
 	 */
-	public byte getColor() {
+	public int getColor() {
 		return step.getColor();
 	}
